test(api): move to directory named api
