+++ workers/tests/integrationtests/api/test_report.py
@@ -10,9 +10,11 @@ import geojson
 from fastapi.testclient import TestClient
 
 from ohsome_quality_analyst.api.api import app
-
-from .api_response_schema import get_general_schema, get_report_feature_schema
-from .utils import oqt_vcr
+from tests.integrationtests.api.response_schema import (
+    get_general_schema,
+    get_report_feature_schema,
+)
+from tests.integrationtests.utils import oqt_vcr
 
 ENDPOINT = "/report"
